requirements.txt
flet
pymupdf
Pillow
